demo/requirements.txt
torch
torchvision
tqdm
pillow
